requirements.txt
ccxt
kafka-python
influxdb
